PP_2_3_1 complete

+++ src/java/main/service/UserService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserService {
     void saveUser(User user);
-    User getUserById(int id);
+    User getUserById(long id);
     void updateUser(User user);
-    void deleteUserById(int id);
+    void deleteUserById(long id);
     List<User> getAllUsers();
 }
